refactor: organization of components and pages into folders
