+++ src/components/SearchBar/SearchBar.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import imgLogo from "../assets/images/logo.jpg";
+import imgLogo from "../../assets/images/logo.jpg";
 
 function searchByName(event, value) {
   event.preventDefault();
